Revert "Fix @cacheable self invocation (#2949)"
This reverts commit 26019e8c215b0e4eab3ad6f3ed5ae6fef6eddf0b.
