+++ src/contractTest/java/uk/gov/hmcts/cmc/claimstore/api/provider/GetClaimsContractConfig.java
@@ -19,8 +19,6 @@ import uk.gov.hmcts.cmc.claimstore.rules.ReviewOrderRule;
 import uk.gov.hmcts.cmc.claimstore.services.ClaimService;
 import uk.gov.hmcts.cmc.claimstore.services.IssueDateCalculator;
 import uk.gov.hmcts.cmc.claimstore.services.ResponseDeadlineCalculator;
-import uk.gov.hmcts.cmc.claimstore.services.UserAuthorisationTokenService;
-import uk.gov.hmcts.cmc.claimstore.services.UserInfoService;
 import uk.gov.hmcts.cmc.claimstore.services.UserService;
 import uk.gov.hmcts.cmc.claimstore.services.ccd.CoreCaseDataService;
 import uk.gov.hmcts.cmc.launchdarkly.LaunchDarklyClient;
@@ -67,18 +65,8 @@ public class GetClaimsContractConfig {
     }
 
     @Bean
-    public UserInfoService userInfoService() {
-        return new UserInfoService(idamApi);
-    }
-
-    @Bean
-    public UserAuthorisationTokenService userAuthorisationTokenService() {
-        return new UserAuthorisationTokenService(idamApi, oauth2);
-    }
-
-    @Bean
     public UserService userService() {
-        return new UserService(idamApi, idamCaseworkerProperties, oauth2, userInfoService(), userAuthorisationTokenService());
+        return new UserService(idamApi, idamCaseworkerProperties, oauth2);
     }
 
     @Bean
