+++ src/main/java/uk/gov/hmcts/cmc/claimstore/services/UserService.java
@@ -3,6 +3,7 @@ package uk.gov.hmcts.cmc.claimstore.services;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.cmc.claimstore.config.properties.idam.IdamCaseworker;
 import uk.gov.hmcts.cmc.claimstore.config.properties.idam.IdamCaseworkerProperties;
@@ -18,32 +19,28 @@ public class UserService {
     public static final String AUTHORIZATION_CODE = "authorization_code";
     public static final String CODE = "code";
     public static final String BASIC = "Basic ";
+    public static final String GRANT_TYPE_PASSWORD = "password";
+    public static final String DEFAULT_SCOPE = "openid profile roles";
     private final IdamApi idamApi;
     private final IdamCaseworkerProperties idamCaseworkerProperties;
     private final Oauth2 oauth2;
-    private final UserInfoService userInfoService;
-    private final UserAuthorisationTokenService userAuthorisationTokenService;
     Logger logger = LoggerFactory.getLogger(this.getClass());
 
     @Autowired
     public UserService(
         IdamApi idamApi,
         IdamCaseworkerProperties idamCaseworkerProperties,
-        Oauth2 oauth2,
-        UserInfoService userInfoService,
-        UserAuthorisationTokenService userAuthorisationTokenService
+        Oauth2 oauth2
     ) {
         this.idamApi = idamApi;
         this.idamCaseworkerProperties = idamCaseworkerProperties;
         this.oauth2 = oauth2;
-        this.userInfoService = userInfoService;
-        this.userAuthorisationTokenService = userAuthorisationTokenService;
     }
 
     @LogExecutionTime
     public UserDetails getUserDetails(String authorisation) {
         logger.info("User info invoked");
-        UserInfo userInfo = userInfoService.getUserInfo(authorisation);
+        UserInfo userInfo = getUserInfo(authorisation);
 
         return UserDetails.builder()
             .id(userInfo.getUid())
@@ -61,7 +58,7 @@ public class UserService {
 
     public User authenticateUser(String username, String password) {
 
-        String authorisation = userAuthorisationTokenService.getAuthorisationToken(username, password);
+        String authorisation = getAuthorisationToken(username, password);
         UserDetails userDetails = getUserDetails(authorisation);
         return new User(authorisation, userDetails);
     }
@@ -77,8 +74,26 @@ public class UserService {
         return idamApi.generatePin(new GeneratePinRequest(name), authorisation);
     }
 
+    @LogExecutionTime
+    @Cacheable(value = "userOIDTokenCache")
+    public String getAuthorisationToken(String username, String password) {
+        logger.info("IDAM /o/token invoked.");
+        TokenExchangeResponse tokenExchangeResponse = idamApi.exchangeToken(
+            oauth2.getClientId(),
+            oauth2.getClientSecret(),
+            oauth2.getRedirectUrl(),
+            GRANT_TYPE_PASSWORD,
+            username,
+            password,
+            DEFAULT_SCOPE);
+        return BEARER + tokenExchangeResponse.getAccessToken();
+    }
+
+    @LogExecutionTime
+    @Cacheable(value = "userInfoCache")
     public UserInfo getUserInfo(String bearerToken) {
-        return userInfoService.getUserInfo(bearerToken);
+        logger.info("IDAM /o/userinfo invoked");
+        return idamApi.retrieveUserInfo(bearerToken);
     }
 
     public User authenticateUserForTests(String username, String password) {
@@ -96,7 +111,7 @@ public class UserService {
             oauth2.getClientId(),
             oauth2.getRedirectUrl()
         );
-        logger.info("IDAM /o/token invoked for tests.");
+        logger.info("IDAM /o/token invoked.");
         TokenExchangeResponse tokenExchangeResponse = idamApi.exchangeTokenForTests(
             authenticateUserResponse.getCode(),
             AUTHORIZATION_CODE,
