--- src/test/java/uk/gov/hmcts/cmc/claimstore/services/UserInfoServiceTest.java
@@ -1,26 +1,0 @@
-package uk.gov.hmcts.cmc.claimstore.services;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-import uk.gov.hmcts.cmc.claimstore.requests.idam.IdamApi;
-
-import static org.mockito.Mockito.verify;
-
-@ExtendWith(MockitoExtension.class)
-class UserInfoServiceTest {
-    private static final String AUTHORISATION = "Bearer I am a valid token";
-
-    @Mock
-    private IdamApi idamApi;
-    @InjectMocks
-    private UserInfoService userInfoService;
-
-    @Test
-    void findsUserInfoForAuthToken() {
-        userInfoService.getUserInfo(AUTHORISATION);
-        verify(idamApi).retrieveUserInfo(AUTHORISATION);
-    }
-}
